+++ src/main/java/org/openpatch/scratch/extensions/hitbox/Hitbox.java
@@ -1,19 +1,15 @@
 package org.openpatch.scratch.extensions.hitbox;
 
-import java.awt.Point;
-import java.awt.Polygon;
-import java.awt.Shape;
-import java.awt.geom.AffineTransform;
-import java.awt.geom.Area;
-import java.awt.geom.PathIterator;
-import java.awt.geom.Rectangle2D;
 import org.openpatch.scratch.Window;
-import org.openpatch.scratch.extensions.math.Utils;
-import processing.core.PConstants;
+import org.openpatch.scratch.extensions.shape.Shape;
+import org.openpatch.scratch.extensions.shape.Bounds;
+import org.openpatch.scratch.extensions.shape.Polygon;
+
 import processing.core.PGraphics;
 
 /**
- * Represents a hitbox with a shape that can be transformed by scaling, translating, and rotating
+ * Represents a hitbox with a shape that can be transformed by scaling,
+ * translating, and rotating
  * it.
  */
 public class Hitbox {
@@ -36,9 +32,9 @@ public class Hitbox {
    * @param xPoints The x-coordinates of the points.
    * @param yPoints The y-coordinates of the points.
    */
-  public Hitbox(int[] xPoints, int[] yPoints) {
-    this.originalShape = new Polygon(xPoints, yPoints, xPoints.length);
-    this.shape = new Polygon(xPoints, yPoints, xPoints.length);
+  public Hitbox(double[] xPoints, double[] yPoints) {
+    this.originalShape = new Polygon(xPoints, yPoints);
+    this.shape = new Polygon(xPoints, yPoints);
   }
 
   /**
@@ -53,12 +49,13 @@ public class Hitbox {
   /**
    * Transforms the shape by scaling, translating, and rotating it.
    *
-   * @param degrees The angle in degrees to rotate the shape.
-   * @param originX The x-coordinate of the rotation origin.
-   * @param originY The y-coordinate of the rotation origin.
+   * @param degrees    The angle in degrees to rotate the shape.
+   * @param originX    The x-coordinate of the rotation origin.
+   * @param originY    The y-coordinate of the rotation origin.
    * @param translateX The x-coordinate to translate the shape.
    * @param translateY The y-coordinate to translate the shape.
-   * @param size The size to scale the shape, where 100.0 represents the original size.
+   * @param size       The size to scale the shape, where 100.0 represents the
+   *                   original size.
    */
   public void translateAndRotateAndResize(
       double degrees,
@@ -68,17 +65,9 @@ public class Hitbox {
       double translateY,
       double size) {
 
-    AffineTransform tx = new AffineTransform();
-    tx.scale(size / 100.0, size / 100.0);
-    this.shape = tx.createTransformedShape(this.originalShape);
-
-    tx = new AffineTransform();
-    tx.translate(translateX, translateY);
-    this.shape = tx.createTransformedShape(this.shape);
-
-    tx = new AffineTransform();
-    tx.rotate(Utils.degreesToRadians(degrees), originX, originY);
-    this.shape = tx.createTransformedShape(this.shape);
+    this.shape = this.originalShape.scale(size / 100.0, size / 100.0);
+    this.shape = this.shape.translate(translateX, translateY);
+    this.shape = this.shape.rotate(degrees, originX, originY);
   }
 
   private void drawDebug(PGraphics buffer, double r, double g, double b) {
@@ -104,14 +93,15 @@ public class Hitbox {
    *
    * @param x The x-coordinate of the point.
    * @param y The y-coordinate of the point.
-   * @return {@code true} if the hitbox contains the point, {@code false} otherwise.
+   * @return {@code true} if the hitbox contains the point, {@code false}
+   *         otherwise.
    */
   public boolean contains(double x, double y) {
-    return this.shape.contains(new Point((int) Math.round(x), (int) Math.round(y)));
+    return this.shape.contains(x, y);
   }
 
-  public Rectangle2D getBounds() {
-    return this.getShape().getBounds2D();
+  public Bounds getBounds() {
+    return this.shape.getBounds();
   }
 
   /**
@@ -121,14 +111,9 @@ public class Hitbox {
    * @return {@code true} if the hitboxes intersect, {@code false} otherwise.
    */
   public boolean intersects(Hitbox hitbox) {
-    if (this.getShape().getBounds2D().intersects(hitbox.getShape().getBounds2D())) {
+    if (this.getShape().getBounds().intersects(hitbox.getShape().getBounds())) {
 
-      Area a = new Area(this.getShape());
-      Area b = new Area(hitbox.getShape());
-
-      a.intersect(b);
-
-      return !a.isEmpty();
+      return this.getShape().intersects(hitbox.getShape());
     }
     return false;
   }
@@ -142,30 +127,6 @@ public class Hitbox {
     if (shape == null) {
       return;
     }
-    PathIterator path = this.shape.getPathIterator(null, 1);
-    final float coord[] = new float[6];
-    boolean began = false;
-    while (!path.isDone()) {
-      final int code = path.currentSegment(coord);
-      if (code == PathIterator.SEG_MOVETO) {
-        buffer.beginShape();
-        began = true;
-        buffer.vertex(coord[0], coord[1]);
-      } else if (code == PathIterator.SEG_LINETO) {
-        buffer.vertex(coord[0], coord[1]);
-      } else if (code == PathIterator.SEG_CLOSE) {
-        buffer.endShape(PConstants.CLOSE);
-        began = false;
-      } else if (code == PathIterator.SEG_CUBICTO) {
-        buffer.vertex(coord[0], coord[1]);
-        buffer.bezierVertex(coord[0], coord[1], coord[2], coord[3], coord[4], coord[5]);
-      } else if (code == PathIterator.SEG_QUADTO) {
-        buffer.quadraticVertex(coord[0], coord[1], coord[2], coord[3]);
-      }
-      path.next();
-    }
-    if (began) {
-      buffer.endShape();
-    }
+    this.shape.draw(buffer);
   }
 }