+++ src/main/java/org/openpatch/scratch/extensions/tiled/Polyline.java
@@ -3,10 +3,12 @@ package org.openpatch.scratch.extensions.tiled;
 import java.util.List;
 
 /**
- * The Polyline class represents a polyline object in Tiled. It includes a list of points that make
+ * The Polyline class represents a polyline object in Tiled. It includes a list
+ * of points that make
  * up the polyline.
  *
- * <p>Example usage:
+ * <p>
+ * Example usage:
  *
  * <pre>{@code
  * Polyline p = new Polyline();
@@ -15,7 +17,7 @@ import java.util.List;
  * p.points.add(new Point(10, 10));
  * }</pre>
  */
-public class Polyline {
+class Polyline {
   /** The list of points that make up the polyline. */
   public List<Point> points;
 }